Move plotting code and add boilerplate for plotly
