+++ src/gemdat/plots/__init__.py
@@ -21,6 +21,7 @@ from .plotly import (
     msd_per_element,
     plot_3d,
     radial_distribution,
+    radial_distribution_between_species,
     rectilinear,
     shape,
     vibrational_amplitudes,
@@ -44,6 +45,7 @@ __all__ = [
     'msd_per_element',
     'plot_3d',
     'radial_distribution',
+    'radial_distribution_between_species',
     'rectilinear',
     'shape',
     'vibrational_amplitudes',
