+++ src/gemdat/plots/plotly/__init__.py
@@ -17,6 +17,7 @@ from ._jumps_vs_time import jumps_vs_time
 from ._msd_per_element import msd_per_element
 from ._plot3d import plot_3d
 from ._radial_distribution import radial_distribution
+from ._radial_distribution_between_species import radial_distribution_between_species
 from ._rectilinear import rectilinear
 from ._shape import shape
 from ._vibrational_amplitudes import vibrational_amplitudes
@@ -32,12 +33,12 @@ __all__ = [
     'energy_along_path',
     'frequency_vs_occurence',
     'jumps_3d',
-    'jumps_3d_animation',
     'jumps_vs_distance',
     'jumps_vs_time',
     'msd_per_element',
     'plot_3d',
     'radial_distribution',
+    'radial_distribution_between_species',
     'rectilinear',
     'shape',
     'vibrational_amplitudes',
